Disposable tracking solved with volatile enlistment, loads of sender tests in memory added
Some line ending issues fixed
